Alternate way to structure create-dagster - two CLI modules that depend on a shared core moduel
Summary:
Wanted to send this out quick in contrast to https://github.com/dagster-io/dagster/pull/30094 as an alternate way that we could structure this with zero code duplication:

- Shared module used by both with no entrypoints (currently using dagster_dg for this, we could call it dagster_dg_core or dagster_dg_shared if we wanted)
- Full dg cli goes in dagster-dg-cli, which we want to make anyway
- pared down create-dagster CLI goes in create_dagster, exposes the exact same aliased scaffold project and scaffold workspace commands

What I like about this is we can make changes to the commands in one place and count on them staying in sync. What I don't like about it is that the pypi module structure is potentially confusing.

> Insert changelog entry or delete this section.
