+++ pyright/master/requirements-pinned.txt
@@ -109,6 +109,7 @@ cssselect==1.3.0
 cssutils==2.11.1
 cycler==0.12.1
 daff==1.3.46
+-e python_modules/libraries/create-dagster
 -e python_modules/dagit
 -e python_modules/dagster
 -e python_modules/libraries/dagster-airbyte
