+++ python_modules/libraries/create-dagster/MANIFEST.in
@@ -1,2 +1,3 @@
 include LICENSE
 include create_dagster/py.typed
+recursive-include create_dagster/templates *
