+++ python_modules/libraries/create-dagster/create_dagster/templates/PROJECT_NAME_PLACEHOLDER/.gitignore.jinja
