+++ python_modules/libraries/create-dagster/create_dagster/templates/PROJECT_NAME_PLACEHOLDER/src/PROJECT_NAME_PLACEHOLDER/definitions.py.jinja
