+++ python_modules/libraries/create-dagster/tox.ini
@@ -12,9 +12,14 @@ passenv =
 install_command = uv pip install {opts} {packages}
 deps =
   pytest
-  -e ../dagster-dg
+  -e ../dagster-dg[test]
+  -e ../dagster-shared
+  -e ../dagster-cloud-cli
   -e ../dagster-cloud-cli
   -e ../dagster-shared
+  -e ../../../python_modules/dagster-graphql
+  -e ../../../python_modules/dagster-pipes
+  -e ../../../python_modules/dagster
   -e .
 allowlist_externals =
   /bin/bash
