+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/plus_tests/test_scaffold_plus_commands.py
@@ -253,9 +253,9 @@ def setup_populated_git_workspace():
             ["git", "remote", "add", "origin", "git@github.com:hooli/example-repo.git"],
             check=True,
         )
-        runner.invoke("scaffold", "project", "--no-uv-sync", "foo")
-        runner.invoke("scaffold", "project", "--no-uv-sync", "bar")
-        runner.invoke("scaffold", "project", "--no-uv-sync", "baz")
+        runner.invoke_create_dagster("project", "--no-uv-sync", "foo")
+        runner.invoke_create_dagster("project", "--no-uv-sync", "bar")
+        runner.invoke_create_dagster("project", "--no-uv-sync", "baz")
         yield runner
 
 
@@ -341,11 +341,11 @@ def test_scaffold_github_actions_command_no_git_root_serverless(
         tempfile.TemporaryDirectory() as temp_dir,
         pushd(temp_dir),
     ):
-        runner.invoke("scaffold", "workspace", "dagster-workspace")
+        runner.invoke_create_dagster("workspace", "dagster-workspace")
         with pushd("dagster-workspace"):
-            runner.invoke("scaffold", "project", "--uv-sync", "foo")
-            runner.invoke("scaffold", "project", "--uv-sync", "bar")
-            runner.invoke("scaffold", "project", "--uv-sync", "baz")
+            runner.invoke_create_dagster("project", "--uv-sync", "foo")
+            runner.invoke_create_dagster("project", "--uv-sync", "bar")
+            runner.invoke_create_dagster("project", "--uv-sync", "baz")
 
             result = runner.invoke("scaffold", "build-artifacts")
             assert result.exit_code == 0, result.output + " " + str(result.exception)
@@ -502,7 +502,7 @@ def test_scaffold_github_actions_git_root_above_workspace(
         ProxyRunner.test(use_fixed_test_components=True) as runner,
         isolated_example_workspace(runner),
     ):
-        runner.invoke("scaffold", "project", "--uv-sync", "foo")
+        runner.invoke_create_dagster("project", "--uv-sync", "foo")
 
         # Setup git workspace in parent directory
         subprocess.run(["git", "init"], check=True, cwd=Path.cwd().parent)
@@ -544,7 +544,7 @@ def test_scaffold_github_actions_git_root_above_project(
         tempfile.TemporaryDirectory() as temp_dir,
         pushd(temp_dir),
     ):
-        runner.invoke("scaffold", "project", "--uv-sync", "foo")
+        runner.invoke_create_dagster("project", "--uv-sync", "foo")
         with pushd("foo"):
             # Setup git workspace in parent directory
             subprocess.run(["git", "init"], check=True, cwd=Path.cwd().parent)
