+++ python_modules/libraries/dagster-dg/tox.ini
@@ -20,6 +20,7 @@ deps =
   -e ../../../python_modules/dagster-pipes
   -e ../../../python_modules/libraries/dagster-shared
   -e ../../../python_modules/libraries/dagster-cloud-cli
+  -e ../../../python_modules/libraries/create-dagster
   plus: -e ../../../python_modules/libraries/dagster-aws
   plus: -e ../../../python_modules/libraries/dagster-k8s
   plus: -e ../../../python_modules/libraries/dagster-docker
