+++ python_modules/libraries/dagster-dlt/tox.ini
@@ -14,6 +14,7 @@ deps =
   -e ../../dagster[test]
   -e ../../dagster-test
   -e ../../libraries/dagster-dg[test]
+  -e ../../libraries/create-dagster
   -e ../../dagster-pipes
   -e ../../dagster-graphql
   -e ../dagster-shared
