+++ scripts/install_dev_python_modules.py
@@ -61,6 +61,7 @@ def main(
         "python_modules/dagster-webserver",
         "python_modules/dagit",
         "python_modules/libraries/dagster-shared",
+        "python_modules/libraries/create-dagster",
         "python_modules/libraries/dagster-managed-elements",
         "python_modules/libraries/dagster-airbyte",
         "python_modules/libraries/dagster-aws[stubs,test]",
@@ -77,6 +78,7 @@ def main(
         "python_modules/libraries/dagster-deltalake-pandas",
         "python_modules/libraries/dagster-deltalake-polars",
         "python_modules/libraries/dagster-dg",
+        "python_modules/libraries/dagster-dg-cli",
         "python_modules/libraries/dagster-dlt",
         "python_modules/libraries/dagster-docker",
         "python_modules/libraries/dagster-gcp[test, dataproc]",
